JIP-57 Auth Pull request 전 수정     - interface.ts를 type.ts로 변경     - CamelCase로 enum 명 설정
